move subset lb test to extensions
