+++ test/common/upstream/load_balancer_benchmark.cc
@@ -546,88 +546,6 @@ BENCHMARK(benchmarkMaglevLoadBalancerWeighted)
     ->Args({500, 95, 75, 25, 10000})
     ->Unit(::benchmark::kMillisecond);
 
-class SubsetLbTester : public BaseTester {
-public:
-  SubsetLbTester(uint64_t num_hosts, bool single_host_per_subset)
-      : BaseTester(num_hosts, 0, 0, true /* attach metadata */) {
-    envoy::config::cluster::v3::Cluster::LbSubsetConfig subset_config;
-    subset_config.set_fallback_policy(
-        envoy::config::cluster::v3::Cluster::LbSubsetConfig::ANY_ENDPOINT);
-    auto* selector = subset_config.mutable_subset_selectors()->Add();
-    selector->set_single_host_per_subset(single_host_per_subset);
-    *selector->mutable_keys()->Add() = std::string(metadata_key);
-
-    subset_info_ = std::make_unique<LoadBalancerSubsetInfoImpl>(subset_config);
-    auto child_lb_creator = std::make_unique<LegacyChildLoadBalancerCreatorImpl>(
-        LoadBalancerType::Random, absl::nullopt, absl::nullopt, absl::nullopt, absl::nullopt,
-        common_config_);
-    lb_ = std::make_unique<SubsetLoadBalancer>(*subset_info_, std::move(child_lb_creator),
-                                               priority_set_, &local_priority_set_, stats_,
-                                               stats_scope_, runtime_, random_, simTime());
-
-    const HostVector& hosts = priority_set_.getOrCreateHostSet(0).hosts();
-    ASSERT(hosts.size() == num_hosts);
-    orig_hosts_ = std::make_shared<HostVector>(hosts);
-    smaller_hosts_ = std::make_shared<HostVector>(hosts.begin() + 1, hosts.end());
-    ASSERT(smaller_hosts_->size() + 1 == orig_hosts_->size());
-    orig_locality_hosts_ = makeHostsPerLocality({*orig_hosts_});
-    smaller_locality_hosts_ = makeHostsPerLocality({*smaller_hosts_});
-  }
-
-  // Remove a host and add it back.
-  void update() {
-    priority_set_.updateHosts(0,
-                              HostSetImpl::partitionHosts(smaller_hosts_, smaller_locality_hosts_),
-                              nullptr, {}, host_moved_, absl::nullopt);
-    priority_set_.updateHosts(0, HostSetImpl::partitionHosts(orig_hosts_, orig_locality_hosts_),
-                              nullptr, host_moved_, {}, absl::nullopt);
-  }
-
-  std::unique_ptr<LoadBalancerSubsetInfoImpl> subset_info_;
-  std::unique_ptr<SubsetLoadBalancer> lb_;
-  HostVectorConstSharedPtr orig_hosts_;
-  HostVectorConstSharedPtr smaller_hosts_;
-  HostsPerLocalitySharedPtr orig_locality_hosts_;
-  HostsPerLocalitySharedPtr smaller_locality_hosts_;
-  HostVector host_moved_;
-};
-
-void benchmarkSubsetLoadBalancerCreate(::benchmark::State& state) {
-  const bool single_host_per_subset = state.range(0);
-  const uint64_t num_hosts = state.range(1);
-
-  if (benchmark::skipExpensiveBenchmarks() && num_hosts > 100) {
-    state.SkipWithError("Skipping expensive benchmark");
-    return;
-  }
-
-  for (auto _ : state) { // NOLINT: Silences warning about dead store
-    SubsetLbTester tester(num_hosts, single_host_per_subset);
-  }
-}
-
-BENCHMARK(benchmarkSubsetLoadBalancerCreate)
-    ->Ranges({{false, true}, {50, 2500}})
-    ->Unit(::benchmark::kMillisecond);
-
-void benchmarkSubsetLoadBalancerUpdate(::benchmark::State& state) {
-  const bool single_host_per_subset = state.range(0);
-  const uint64_t num_hosts = state.range(1);
-  if (benchmark::skipExpensiveBenchmarks() && num_hosts > 100) {
-    state.SkipWithError("Skipping expensive benchmark");
-    return;
-  }
-
-  SubsetLbTester tester(num_hosts, single_host_per_subset);
-  for (auto _ : state) { // NOLINT: Silences warning about dead store
-    tester.update();
-  }
-}
-
-BENCHMARK(benchmarkSubsetLoadBalancerUpdate)
-    ->Ranges({{false, true}, {50, 2500}})
-    ->Unit(::benchmark::kMillisecond);
-
 } // namespace
 } // namespace Upstream
 } // namespace Envoy
